+++ samples/SATokenSample.py
@@ -1,9 +1,16 @@
 from skyflow.Errors import SkyflowError
-from skyflow.ServiceAccount import generateBearerToken
+from skyflow.ServiceAccount import generateBearerToken, isValid
 
+# cache token for reuse
+bearerToken = ''
+tokenType = ''
+def tokenProvider():
+    if not isValid(bearerToken):
+        bearerToken, tokenType = generateBearerToken('<YOUR_CREDENTIALS_FILE_PATH>')
+    return bearerToken, tokenType
 
 try:
-    accessToken, tokenType = generateBearerToken('<YOUR_CREDENTIALS_FILE_PATH>')
+    accessToken, tokenType = tokenProvider()
     print("Access Token:", accessToken)
     print("Type of token:", tokenType)
 except SkyflowError as e: