+++ samples/detokenizeSample.py
@@ -1,11 +1,13 @@
 from skyflow.Errors import SkyflowError
-from skyflow.ServiceAccount import generateBearerToken
+from skyflow.ServiceAccount import generateBearerToken, isValid
 from skyflow.Vault import Client, Configuration
 
-
+# cache token for reuse
+bearerToken = ''
 def tokenProvider():
-    token, _ = generateBearerToken('<YOUR_CREDENTIALS_FILE_PATH>')
-    return token
+    if not isValid(bearerToken):
+        bearerToken, _ = generateBearerToken('<YOUR_CREDENTIALS_FILE_PATH>')
+    return bearerToken
 
 try:
     config = Configuration('<YOUR_VAULT_ID>', '<YOUR_VAULT_URL>', tokenProvider)