+++ samples/invokeConnectionSample.py
@@ -1,15 +1,18 @@
 from skyflow.Errors import SkyflowError
-from skyflow.ServiceAccount import generateBearerToken
+from skyflow.ServiceAccount import generateBearerToken, isValid
 from skyflow.Vault import Client, Configuration, RequestMethod, ConnectionConfig
 
 '''
 This sample is for generating CVV using Skyflow Connection with a third party integration such as VISA
 '''
 
-
+# cache token for reuse
+bearerToken = ''
 def tokenProvider():
-    token, _ = generateBearerToken('<YOUR_CREDENTIALS_FILE_PATH>')
-    return token
+    if not isValid(bearerToken):
+        bearerToken, _ = generateBearerToken('<YOUR_CREDENTIALS_FILE_PATH>')
+    return bearerToken
+
 
 try:
     config = Configuration('<YOUR_VAULT_ID>', '<YOUR_VAULT_URL>', tokenProvider)