+++ skyflow/Vault/_detokenize.py
@@ -1,6 +1,6 @@
 from skyflow.Errors._skyflowErrors import SkyflowError, SkyflowErrorCodes, SkyflowErrorMessages
 import asyncio
-from aiohttp import ClientSession
+from aiohttp import ClientSession, request
 import json
 from skyflow._utils import InterfaceName
 
@@ -50,7 +50,10 @@ async def sendDetokenizeRequests(data, url, token):
 
 async def post(url, data, headers, session):
     async with session.post(url, data=data, headers=headers, ssl=False) as response:
-        return (await response.read(), response.status)
+        try:
+            return (await response.read(), response.status, response.headers['x-request-id'])
+        except KeyError:
+            return (await response.read(), response.status) 
 
 def createDetokenizeResponseBody(responses):
     result = {
@@ -72,6 +75,8 @@ def createDetokenizeResponseBody(responses):
             temp = {"error": {}}
             temp["error"]["code"] = jsonRes["error"]["http_code"]
             temp["error"]["description"] = jsonRes["error"]["message"]
+            if len(r) >= 2 and r[2] != None:
+                temp["error"]["description"] += ' - Request ID: ' + str(r[2])
             result["errors"].append(temp)
             partial = True
     return result, partial