+++ tests/Vault/testInvokeConnection.py
@@ -149,3 +149,4 @@ class testInvokeConnection(unittest.TestCase):
     #         print(e)
             # self.fail()
 
+